requirements.txt
streamlit
sqlalchemy
psycopg2-binary
Pillow